feat-[166272336]: signout user
- drop user token
- logout user
- authorize user token
- test the logout endpoint

[Finishes #166272336]
