+++ server/helpers/Token.js
@@ -1,6 +1,9 @@
 import jwt from 'jsonwebtoken';
 import { config } from 'dotenv';
 import Response from './Response';
+import models from '../models';
+
+const { User, DroppedToken } = models;
 
 config();
 const tokenSecret = process.env.SECRET || 'secret';
@@ -23,27 +26,56 @@ class Token {
   }
 
   /**
-   * @static
-   * @param {*} req the request object
-   * @param {*} res the response object
-   * @param {*} next middleware
-   * @returns {function} next
-   * @memberof Token
-   */
-  static async verifyToken(req, res, next) {
+  * Get token from req
+  * @param  {object} req - Request object
+  * @returns {string} token
+  * @static
+  */
+  static getToken(req) {
+    const bearerToken = req.headers.authorization;
+    const token = bearerToken.startsWith('Bearer ') && bearerToken.replace('Bearer ', '');
+
+    return token;
+  }
+
+  /**
+  * Handles authorization of users
+  * @async
+  * @param  {object} req - Request object
+  * @param {object} res - Response object
+  * @param {object} next The next middleware
+  * @return {json} Returns json object
+  * @static
+  */
+  static async authorize(req, res, next) {
     try {
-      const { token } = req.headers;
-      if (typeof token === 'undefined') {
-        return Response.error(res, 400, 'No token provided!');
+      const token = await Token.getToken(req);
+      if (!token) return Response.error(res, 400, 'Invalid token supplied: format Bearer <token>');
+
+      const decoded = jwt.verify(token, tokenSecret);
+      const user = await User.findOne({ where: { id: decoded.id } });
+      if (!user) {
+        return Response.error(res, 401, 'Invalid Token Provided');
       }
-      const verifiedToken = await jwt.verify(token, tokenSecret);
-      if (!verifiedToken) {
-        return Response.error(res, 401, 'Token cannot be verified!');
+
+      const droppedToken = await DroppedToken.findOne({
+        where: {
+          token
+        }
+      });
+      if (droppedToken) {
+        return res.status(401).json({
+          status: 401,
+          error: 'This token has been dropped'
+        });
       }
-      req.decoded = verifiedToken;
-      return next();
-    } catch (err) {
-      next(err);
+      req.decoded = decoded;
+      next();
+    } catch (error) {
+      if (error.name === 'JsonWebTokenError') {
+        return Response.error(res, 401, 'Invalid Token');
+      }
+      next(error);
     }
   }
 }
